Make InvalidBitPattern::invalid_bytes field private
Clearly there was an intent to make this a normal closed API
since there was already a (partially untested) impl.

This patch also reorganises code to eliminate early error returns
when the early return is at the end of a function and so can be
trivially represented as an if/else.